to_sql_checkedの型変換をi32に修正

+++ src/packet/types/protocol/ether_type.rs
@@ -57,6 +57,6 @@ impl ToSql for EtherType {
         ty: &Type,
         out: &mut BytesMut,
     ) -> Result<IsNull, Box<dyn Error + Sync + Send>> {
-        (self.0 as i16).to_sql_checked(ty, out)
+        (self.0 as i32).to_sql_checked(ty, out)
     }
 }
